+++ src/Config/GenericConfig.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Wimski\Nominatim\Config;
+
+use Wimski\Nominatim\Contracts\Config\GenericConfigInterface;
+
+class GenericConfig extends AbstractConfig implements GenericConfigInterface
+{
+}