+++ src/Contracts/Config/GenericConfigInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Wimski\Nominatim\Contracts\Config;
+
+interface GenericConfigInterface extends ConfigInterface
+{
+}